Async Study
Watched the videos and read the material and answered the questions to
the best of my ability.

I feel comfortable with the concepts at a high level but will need to
see it in action a bit more to feel like I can write code that is
mindful of the call stack.